refactor(types): merge two files into one
