+++ src/types/index.ts
@@ -7,3 +7,7 @@ export interface Entry extends fs.Stats {
 }
 
 export type EntryItem = string | Entry;
+
+export type Pattern = string;
+export type PatternRe = RegExp;
+export type PatternsGroup = Record<string, Pattern[]>;
