--- src/types/patterns.ts
@@ -1,3 +1,0 @@
-export type Pattern = string;
-export type PatternRe = RegExp;
-export type PatternsGroup = Record<string, Pattern[]>;
